+++ tests/testthat/test-nmbr.R
@@ -6,6 +6,14 @@ test_that("nmbr works for scalar values", {
   expect_equal(nmbr(-0.123456, scale = 100, suffix = "%"), "&minus;12%")
   expect_equal(nmbr(123456.789, big.mark = ".", decimal.mark = ",", accuracy = 0.001),
                "123.456,789")
+  expect_equal(nmbr(-123456.789, big.mark = ".", decimal.mark = ",", accuracy = 0.001, bold = TRUE),
+               "**&minus;123.456,789**")
+  expect_equal(nmbr(-123456.789, big.mark = ".", decimal.mark = ",", accuracy = 0.001,
+                    italic = TRUE),
+               "*&minus;123.456,789*")
+  expect_equal(nmbr(-123456.789, big.mark = ".", decimal.mark = ",", accuracy = 0.001,
+                    italic = TRUE, bold = TRUE),
+               "***&minus;123.456,789***")
   expect_equal(nmbr(numeric()), character())
   expect_equal(nmbr(NA_real_), NA_character_)
   expect_equal(nmbr(NA_real_, na = "NA"), "NA")
@@ -25,20 +33,34 @@ test_that("nmbr works for vector formatting options", {
          big.mark = c("", " ", ","), decimal.mark = c("/", ",", ".")),
     c("0/12", "&minus;1 234,56", "123,456.00")
   )
+  expect_equal(
+    nmbr(c(0.123456, -1234.56, 123456), accuracy = 0.01,
+         big.mark = c("", " ", ","), decimal.mark = c("/", ",", "."),
+         bold = c(TRUE, TRUE, FALSE), italic = c(TRUE, FALSE, TRUE)),
+    c("***0/12***", "**&minus;1 234,56**", "*123,456.00*")
+  )
 
   test_that("prct works as expected", {
     expect_equal(prct(0.123456), nmbr(0.123456, scale = 100, suffix = "%"))
     expect_equal(prct(0.987654, percent = ""), nmbr(0.987654, scale = 100))
+    expect_equal(prct(0.987654, bold = TRUE),
+                 nmbr(0.987654, scale = 100, suffix = "%", bold = TRUE))
+    expect_equal(prct(0.987654, italic = TRUE),
+                 nmbr(0.987654, scale = 100, suffix = "%", italic = TRUE))
   })
 
   test_that("cmma works as expected", {
     expect_equal(cmma(123456), nmbr(123456, big.mark = ","))
     expect_equal(cmma(987654, comma = "< >"), nmbr(987654))
+    expect_equal(cmma(123456, bold = TRUE), nmbr(123456, big.mark = ",", bold = TRUE))
+    expect_equal(cmma(123456, italic = TRUE), nmbr(123456, big.mark = ",", italic = TRUE))
   })
 
   test_that("dllr works as expected", {
     expect_equal(dllr(123456), nmbr(123456, prefix = "$"))
     expect_equal(dllr(987654, dollar = "\u20ac"), nmbr(987654, prefix = "\u20ac"))
+    expect_equal(dllr(123456, bold = TRUE), nmbr(123456, prefix = "$", bold = TRUE))
+    expect_equal(dllr(123456, italic = TRUE), nmbr(123456, prefix = "$", italic = TRUE))
   })
 
   test_that("pval works as expected", {
@@ -57,6 +79,8 @@ test_that("nmbr gives appropriate error messages", {
   expect_error(nmbr(123456, suffix = 1), class = "formattr_error_wrong_type")
   expect_error(nmbr(123456, big.mark = 1), class = "formattr_error_wrong_type")
   expect_error(nmbr(123456, decimal.mark = 1), class = "formattr_error_wrong_type")
+  expect_error(nmbr(123456, bold = 1), class = "formattr_error_wrong_type")
+  expect_error(nmbr(123456, italic = "1"), class = "formattr_error_wrong_type")
   expect_error(nmbr(123456, html = "a"), class = "formattr_error_wrong_type")
   expect_error(nmbr(123456, na = 1), class = "formattr_error_wrong_type")
   expect_error(nmbr(123456, accuracy = 1:2), class = "formattr_error_wrong_length")
@@ -65,6 +89,8 @@ test_that("nmbr gives appropriate error messages", {
   expect_error(nmbr(123456, suffix = rep("%", 2)), class = "formattr_error_wrong_length")
   expect_error(nmbr(123456, big.mark = rep(",", 2)), class = "formattr_error_wrong_length")
   expect_error(nmbr(123456, decimal.mark = rep(".", 2)), class = "formattr_error_wrong_length")
+  expect_error(nmbr(123456, bold = rep(TRUE, 2)), class = "formattr_error_wrong_length")
+  expect_error(nmbr(123456, italic = rep(FALSE, 2)), class = "formattr_error_wrong_length")
   expect_error(pval(c(0.0123, 0.1234), accuracy = c(1, 2)), class = "formattr_error_wrong_type")
   expect_error(pval(0.0123, min_p = "a"), class = "formattr_error_wrong_type")
   expect_error(pval(0.0123, accuracy = 0.01, min_p = 0.001), class = "formattr_error_invalid_min_p")