+++ NEWS.md
@@ -1,8 +1,14 @@
-# formattr 0.1.0.9000
+# formattr 0.2.0.9000
+
+* Added `bold` and `italic` arguments to `nmbr()` and related functions.
+
+# formattr 0.2.0
+
+* Added `create_descriptive_table() `
 
 * Added `pval()`, `sstars()`, and `coef_se()`.
 
-* Added `bold` and `italic` arguments to `nmbr()` and related functions.
+* Added `create_nmbr()`.
 
 # formattr 0.1.0